+++ gneiss-mqtt/src/client/synchronous/threaded/mod.rs
@@ -20,6 +20,7 @@ use std::time::{Duration, Instant};
 use log::{debug, error, info, trace};
 use crate::client::*;
 use crate::client::config::*;
+use crate::client::waiter::*;
 #[cfg(feature="threaded-websockets")]
 use ws_stream::WebsocketStreamWrapper;
 use crate::error::{GneissError, GneissResult};
@@ -1161,15 +1162,116 @@ fn apply_proxy_connect_to_stream<T>(mut stream : T, http_connect_endpoint: Endpo
     }
 }
 
+/// Simple debug type that uses the client listener framework to allow tests to wait for
+/// configurable client event sequences.
+pub struct ThreadedClientEventWaiter {
+    event_count: usize,
 
+    client: SyncClientHandle,
+
+    listener: Option<ListenerHandle>,
+
+    events: Arc<Mutex<Option<Vec<ClientEventRecord>>>>,
+
+    signal: Arc<Condvar>,
+}
+
+impl ThreadedClientEventWaiter {
+
+    fn new_internal(client: SyncClientHandle, config: ClientEventWaiterOptions, event_count: usize) -> Self {
+        let lock = Arc::new(Mutex::new(Some(Vec::new())));
+        let signal = Arc::new(Condvar::new());
+
+        let mut waiter = ThreadedClientEventWaiter {
+            event_count,
+            client: client.clone(),
+            listener: None,
+            events: lock.clone(),
+            signal: signal.clone(),
+        };
+
+        let listener_fn = move |event: Arc<ClientEvent>| {
+            match &config.wait_type {
+                ClientEventWaitType::Type(event_type) => {
+                    if !client_event_matches(&event, *event_type) {
+                        return;
+                    }
+                }
+                ClientEventWaitType::Predicate(event_predicate) => {
+                    if !(*event_predicate)(&event) {
+                        return;
+                    }
+                }
+            }
+
+            let event_record = ClientEventRecord {
+                event: event.clone(),
+                timestamp: Instant::now(),
+            };
+
+            let mut events_guard = lock.lock().unwrap();
+            let events_option = events_guard.as_mut();
+            if let Some(events) = events_option {
+                events.push(event_record);
+
+                if events.len() >= event_count {
+                    signal.notify_all();
+                }
+            }
+        };
+
+        waiter.listener = Some(client.add_event_listener(Arc::new(listener_fn)).unwrap());
+        waiter
+    }
+
+    /// Creates a new ClientEventWaiter instance from full configuration
+    #[cfg(feature = "testing")]
+    pub fn new(client: SyncClientHandle, config: ClientEventWaiterOptions, event_count: usize) -> Self {
+        Self::new_internal(client, config, event_count)
+    }
+
+    /// Creates a new ClientEventWaiter instance that will wait for a single occurrence of a single event type
+    pub fn new_single(client: SyncClientHandle, event_type: ClientEventType) -> Self {
+        let config = ClientEventWaiterOptions {
+            wait_type: ClientEventWaitType::Type(event_type),
+        };
+
+        Self::new_internal(client, config, 1)
+    }
+
+    /// Waits for the configured event(s) and returns a result with them
+    pub fn wait(self) -> GneissResult<Vec<ClientEventRecord>> {
+        let mut current_events_option = self.events.lock().unwrap();
+        loop {
+            match &*current_events_option {
+                Some(current_events) => {
+                    if current_events.len() >= self.event_count {
+                        return Ok(current_events_option.take().unwrap());
+                    }
+                }
+                None => {
+                    return Err(GneissError::new_other_error("Client event waiter result already taken"));
+                }
+            }
+
+            current_events_option = self.signal.wait(current_events_option).unwrap();
+        }
+    }
+}
+
+impl Drop for ThreadedClientEventWaiter {
+    fn drop(&mut self) {
+        let listener_handler = self.listener.take().unwrap();
+
+        let _ = self.client.remove_event_listener(listener_handler);
+    }
+}
 
 
 #[cfg(feature = "testing")]
 pub(crate) mod testing {
     use crate::client::synchronous::*;
     use crate::testing::integration::*;
-    use crate::testing::waiter::*;
-    use crate::testing::waiter::synchronous::*;
     use super::*;
 
     fn threaded_connect_disconnect_test(builder: ClientBuilder, sync_options: SyncClientOptions, client_options: ThreadedClientOptions) -> GneissResult<()> {
@@ -1363,7 +1465,7 @@ pub(crate) mod testing {
 
     fn connection_failure_test(builder : ClientBuilder, sync_options: SyncClientOptions, threaded_options: ThreadedClientOptions) -> GneissResult<()> {
         let client = builder.build_threaded(sync_options, threaded_options).unwrap();
-        let connection_failure_waiter = SyncClientEventWaiter::new_single(client.clone(), ClientEventType::ConnectionFailure);
+        let connection_failure_waiter = ThreadedClientEventWaiter::new_single(client.clone(), ClientEventType::ConnectionFailure);
 
         client.start(None)?;
